Code refactor to use new api endpoint for pinging service.
