+++ src/SFA.DAS.EmployerFinance.Api/SFA.DAS.EmployerFinance.Api.csproj
@@ -20,8 +20,4 @@
       <ProjectReference Include="..\SFA.DAS.EmployerFinance\SFA.DAS.EmployerFinance.csproj" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="Controllers" />
-    </ItemGroup>
-
 </Project>
